# packages/dashboard/openapi.yaml
openapi: 3.1.0
info:
  title: Aurora Dashboard API
  version: 0.3.0
  description: >
    Admin-plane API for aurora.local. Same-origin, cookie-authenticated
    (session cookie set by POST /auth/login). All paths below are served
    under the `/api` prefix by the Spring Boot backend; the Vite dev
    server proxies `/api` to :8090.


    This spec is hand-authored from the typed frontend `src/api/*`
    modules and is the source of truth the MSW mock layer answers to.
    Field casing is deliberately mixed: most bodies are camelCase, but a
    handful of endpoints emit snake_case verbatim from the backend
    (onboarding draft/plan/launch, services, containers logs, security
    dismissals, audit). Casing here matches the wire, not a convention.
servers:
  - url: /api
    description: Same-origin, behind Caddy / Vite proxy.
...
security:
  - sessionCookie: []
paths:
...
  /packages/{name}/resources:
    get:
      tags: [packages]
      summary: Memory and CPU ceilings, and live usage
      parameters: [{ $ref: '#/components/parameters/PkgName' }]
      responses:
        '200':
          description: Resource state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PackageResources' }
    put:
      tags: [packages]
      summary: Override the ceilings
      description: >
        Null on either field clears the override and restores the
        manifest default. New limits apply the next time the package
        restarts.
      parameters: [{ $ref: '#/components/parameters/PkgName' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [memLimitMb, cpus]
              properties:
                memLimitMb: { type: [integer, 'null'] }
                cpus: { type: [number, 'null'] }
      responses:
        '200':
          description: The saved state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PackageResources' }

  # ─────────────────────── previously undocumented ────────────────────
  # Endpoints that existed in the backend before anything checked the
  # controllers against this file. Found by OpenApiConformanceTest.
...
  /users:
    get:
      tags: [users]
      summary: List users
      responses:
        '200':
          description: Users.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
    post:
      tags: [users]
      summary: Create a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, role, password]
              properties:
                username: { type: string }
                role: { $ref: '#/components/schemas/UserRole' }
                password: { type: string, minLength: 8 }
      responses:
        '201':
          description: Created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '409': { description: Username already taken. }
  /users/{id}:
    get:
      tags: [users]
      summary: One user
      parameters: [{ $ref: '#/components/parameters/UserId' }]
      responses:
        '200':
          description: The user.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404': { description: Unknown user. }
    put:
      tags: [users]
      summary: Update a user's role or password
      parameters: [{ $ref: '#/components/parameters/UserId' }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { $ref: '#/components/schemas/UserRole' }
      responses:
        '200':
          description: Updated.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '404': { description: Unknown user. }
    delete:
      tags: [users]
      summary: Remove a user
      parameters: [{ $ref: '#/components/parameters/UserId' }]
      responses:
        '204': { description: Removed. }
        '409': { description: Cannot remove your own account. }

  # ─────────────────────────────── vpn ────────────────────────────────
  # Aurora's own inbound VPN server for remote access into the LAN.
  # NOT packages/privacy's Gluetun sidecar (an outbound client tunnel
  # for the media stack) — see docs/VPN_PAGE_DESIGN.md.
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: AURORA_SESSION
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema: { type: string }
    PkgName:
      name: name
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    PackageResources:
      type: object
      description: >
        Per-container ceilings for one package. A home box has no spare
        capacity and usually no swap, so one runaway container takes the
        whole machine down; a cap means the kernel kills the offender
        rather than the OOM killer choosing at random.
      required: [package, defaultMemLimitMb, defaultCpus, memLimitMb, cpus, memUsedMb, cpuPct]
      properties:
        package: { type: string }
        defaultMemLimitMb:
          type: [integer, 'null']
          description: From the manifest. Null means the package ships uncapped.
        defaultCpus: { type: [number, 'null'] }
        memLimitMb:
          type: [integer, 'null']
          description: Operator override. Null means the manifest default applies.
        cpus: { type: [number, 'null'] }
        memUsedMb: { type: [integer, 'null'] }
        cpuPct: { type: [number, 'null'] }
...
    UserRole:
      type: string
      description: >
        Phase D's vocabulary. `user` is a regular authenticated identity
        and `guest` is read-mostly; the DDL default is `user` so a caller
        that forgets to specify a role cannot accidentally create an
        admin. Enforced by SQLite triggers, not just application code.
      enum: [admin, user, guest]
    User:
      type: object
      required: [id, username, role, tz, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
        tz: { type: string }
        createdAt: { type: string, format: date-time }